+++ src/contour/shell-integration.zsh
@@ -2,7 +2,7 @@
 #
 #/**
 # * This file is part of the "contour" project
-# *   Copyright (c) 2019-2020 Christian Parpart <christian@parpart.family>
+# *   Copyright (c) 2019-2021 Christian Parpart <christian@parpart.family>
 # *
 # * Licensed under the Apache License, Version 2.0 (the "License");
 # * you may not use this file except in compliance with the License.